pyproject.toml
[tool.poetry]
name = "skepsi_backend"
version = "0.1.0"
description = ""
authors = ["finnianmacken <finn@minerva.kgi.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
rope = "0.18.0"
aniso8601 = "8.1.1"
apollo-client = "2.1.2"
appdirs = "1.4.4"
appnope = "0.1.2"
argon2-cffi = "20.1.0"
asgiref = "3.3.1"
astroid = "2.5.2"
autopep8 = "1.5.5"
backcall = "0.2.0"
bcrypt = "3.2.0"
beautifulsoup4 = "4.9.3"
CacheControl = "0.12.6"
cachy = "0.3.0"
certifi = "2021.5.30"
cffi = "1.14.5"
chardet = "4.0.0"
cleo = "0.8.1"
clikit = "0.6.2"
crashtest = "0.3.1"
cryptography = "3.4.7"
pyls-mypy = "^0.1.8"
decorator = "4.4.2"
defusedxml = "0.7.1"
distlib = "0.3.1"
Django = "3.1.7"
django-bootstrap = "0.2.4"
django-bootstrap4 = "2.3.1"
django-braces = "1.14.0"
django-cors-headers = "3.7.0"
django-debug-toolbar = "3.2"
django-filter = "2.4.0"
django-graphql-jwt = "0.3.2"
djangorestframework = "3.12.4"
djangorestframework-queryfields = "1.0.0"
drf-jwt = "1.19.1"
ecdsa = "0.17.0"
Faker = "6.6.2"
filelock = "3.0.12"
flake8 = "3.8.4"
graphene = "3.0b7"
graphene-django = "3.0.0b7"
graphql-core = "3.1.5"
graphql-relay = "3.1.0"
html5lib = "1.1"
idna = "2.10"
ipykernel = "5.5.0"
ipython = "7.21.0"
ipython_genutils = "0.2.0"
isort = "5.8.0"
jedi = "0.17.2"
jose = "1.0.0"
jsonpickle = "2.0.0"
jupyter-client = "6.1.12"
jupyter-core = "4.7.1"
keyring = "21.8.0"
lazy-object-proxy = "1.6.0"
lockfile = "0.12.2"
mccabe = "0.6.1"
misaka = "2.1.1"
msgpack = "1.0.2"
mypy = "0.812"
mypy-extensions = "0.4.3"
oauthlib = "3.1.1"
packaging = "20.9"
parso = "0.7.1"
pastel = "0.2.1"
pexpect = "4.8.0"
pickleshare = "0.7.5"
Pillow = "8.1.2"
pkginfo = "1.7.0"
pluggy = "0.13.1"
poetry = "1.1.7"
poetry-core = "1.0.3"
promise = "2.3"
prompt-toolkit = "3.0.17"
ptyprocess = "0.7.0"
pyasn1 = "0.4.8"
pycodestyle = "2.6.0"
pycparser = "2.20"
pydocstyle = "6.0.0"
pyflakes = "2.2.0"
Pygments = "2.8.1"
PyJWT = "2.1.0"
pylev = "1.4.0"
pylint = "2.7.4"
pyparsing = "2.4.7"
python-dateutil = "2.8.1"
python-jose = "3.3.0"
python-jsonrpc-server = "0.4.0"
python-language-server = "0.36.2"
python3-openid = "3.2.0"
pytz = "2021.1"
pyzmq = "22.0.3"
requests = "2.25.1"
requests-oauthlib = "1.3.0"
requests-toolbelt = "0.9.1"
rsa = "4.7.2"
Rx = "1.6.1"
shellingham = "1.4.0"
singledispatch = "3.6.2"
six = "1.15.0"
snowballstemmer = "2.1.0"
social-auth-app-django = "4.0.0"
social-auth-core = "4.1.0"
soupsieve = "2.2.1"
sqlparse = "0.4.1"
text-unidecode = "1.3"
toml = "0.10.2"
tomlkit = "0.7.2"
tornado = "6.1"
traitlets = "5.0.5"
typed-ast = "1.4.2"
typing-extensions = "3.7.4.3"
ujson = "4.0.2"
urllib3 = "1.26.5"
virtualenv = "20.4.3"
wcwidth = "0.2.5"
webencodings = "0.5.1"
wrapt = "1.12.1"
yapf = "0.31.0"
psycopg2-binary = "^2.9.1"
dj-database-url = "^0.5.0"
whitenoise = "^5.2.0"
django-heroku = "^0.3.1"
python-dotenv = "^0.18.0"
gunicorn = "^20.1.0"
django-environ = "^0.4.5"
transformers = {extras = ["torch"], version = "^4.9.1"}
django-background-tasks = "^1.2.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
